requirements.txt
wikipedia-api
transformers
torch
faiss-cpu
numpy
openai
fastapi
uvicorn